Circle - User - Follow Relationship
Each circle has one "person of interest" (POI) and circle members (supposed to include POI). The follow mechanism is done by making POI follow each circle members. The action stream for each user U is constructed using the POI's "user stream" and excluding the actions in which U was a part as the actor.